requirements.txt
certifi==2019.3.9
chardet==3.0.4
future==0.17.1
gym==0.12.1
gym-retro==0.7.0
idna==2.8
neat-python==0.92
numpy==1.16.3
opencv-python==4.1.0.25
pyglet==1.3.2
requests==2.21.0
scipy==1.2.1
six==1.12.0
urllib3==1.24.2
